# docs/openapi-client.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Guardian Client API",
    "description": "Client-facing Guardian HTTP API consumed by SDKs and packages.",
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0"
    },
    "version": "0.17.0-rc.3"
  },
  "paths": {
...
    "/configure": {
      "post": {
        "tags": [
          "client"
        ],
        "summary": "Configure (register) an account with its authorization set and\ninitial state. Requires the signed `x-pubkey` / `x-signature` /\n`x-timestamp` auth headers.",
        "operationId": "configure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigureResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or replay rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the configured size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-pubkey": [],
            "x-signature": [],
            "x-timestamp": []
          }
        ]
      }
    },
    "/delta": {
      "get": {
        "tags": [
          "client"
        ],
        "summary": "Fetch the delta for an account at a specific nonce.",
        "operationId": "get_delta",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nonce",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delta found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeltaObject"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or replay rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Delta not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the configured size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-pubkey": [],
            "x-signature": [],
            "x-timestamp": []
          }
        ]
      },
      "post": {
        "tags": [
          "client"
        ],
        "summary": "Push a signed state delta for a single-key account. The request\nbody is the JSON-encoded [`DeltaObject`] to commit.",
        "operationId": "push_delta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeltaObject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Delta accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeltaObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid delta payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or replay rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicting pending delta/proposal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the configured size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-pubkey": [],
            "x-signature": [],
            "x-timestamp": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorMeta": {
        "type": "object",
        "description": "Structured machine-readable side-data on the error wire object\n(feature `009-human-readable-errors`). `retryable` is always present;\nthe rest appear only for the codes that carry them.",
        "required": [
          "retryable"
        ],
        "properties": {
          "missing_permissions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Lex-sorted permissions the operator lacks. Present only for\n`GUARDIAN_INSUFFICIENT_OPERATOR_PERMISSION`."
          },
          "paused_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 3339 pause timestamp. Present only for `GUARDIAN_ACCOUNT_PAUSED`."
          },
          "paused_reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pause reason. Present only for `GUARDIAN_ACCOUNT_PAUSED`."
          },
          "released_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 3339 timestamp of the guardian-switch release. Present only\nfor `GUARDIAN_ACCOUNT_RELEASED`."
          },
          "retry_after_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Seconds to wait before retrying. Present only for\n`rate_limit_exceeded` (the `Retry-After` header carries the same value).",
            "minimum": 0
          },
          "retryable": {
            "type": "boolean",
            "description": "Whether retrying the same request could plausibly succeed."
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "Wire shape of a Guardian error response body: `{ code, message, meta }`\n(feature `009-human-readable-errors`). Mirrors the object produced by\n[`crate::error::GuardianError`]'s `IntoResponse` impl and carried\nidentically on the gRPC `Status.details`. The legacy `success`/`error`\nfields are gone; the diagnostic detail is logged server-side only.\nDocumented as the body of every non-2xx response.",
        "required": [
          "code",
          "message",
          "meta"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code (e.g. `account_not_found`). The\nclient branch + i18n key. Branch on this, never on `message`."
          },
          "message": {
            "type": "string",
            "description": "Short, user-safe message; safe to display verbatim. Wording is not\npart of the stable contract."
          },
          "meta": {
            "$ref": "#/components/schemas/ApiErrorMeta",
            "description": "Structured machine-readable side-data."
          }
        }
      },
      "AssetKind": {
        "type": "string",
        "enum": [
          "fungible",
          "non_fungible"
        ]
      },
      "AssetSummary": {
        "type": "object",
        "required": [
          "asset_id",
          "kind"
        ],
        "properties": {
          "amount": {
            "type": [
              "string",
              "null"
            ],
            "description": "Signed decimal magnitude (e.g., `\"+100\"`, `\"-50\"`) for fungible\nholdings. Absent for non-fungible holdings where the detail\nview uses `added` / `removed` lists instead."
          },
          "asset_id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AssetKind"
          }
        }
      },
      "Auth": {
        "oneOf": [
          {
            "type": "object",
            "description": "Miden Falcon RPO signature scheme",
            "required": [
              "MidenFalconRpo"
            ],
            "properties": {
              "MidenFalconRpo": {
                "type": "object",
                "description": "Miden Falcon RPO signature scheme",
                "required": [
                  "cosigner_commitments"
                ],
                "properties": {
                  "cosigner_commitments": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "description": "Miden ECDSA secp256k1 signature scheme",
            "required": [
              "MidenEcdsa"
            ],
            "properties": {
              "MidenEcdsa": {
                "type": "object",
                "description": "Miden ECDSA secp256k1 signature scheme",
                "required": [
                  "cosigner_commitments"
                ],
                "properties": {
                  "cosigner_commitments": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "description": "EVM ECDSA account signer snapshot",
            "required": [
              "EvmEcdsa"
            ],
            "properties": {
              "EvmEcdsa": {
                "type": "object",
                "description": "EVM ECDSA account signer snapshot",
                "required": [
                  "signers"
                ],
                "properties": {
                  "signers": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "description": "Authentication and authorization handler\nDefines which signature scheme to use and handles verification\nEach variant contains auth-specific authorization data"
      },
      "ConfigureRequest": {
        "type": "object",
        "required": [
          "account_id",
          "auth",
          "initial_state"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "auth": {
            "$ref": "#/components/schemas/Auth"
          },
          "initial_state": {
            "type": "object",
            "description": "Opaque, schema-free JSON blob describing the initial account state."
          },
          "network_config": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NetworkConfig"
              }
            ]
          }
        }
      },
      "ConfigureResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "ack_commitment": {
            "type": [
              "string",
              "null"
            ]
          },
          "ack_pubkey": {
            "type": [
              "string",
              "null"
            ]
          },
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "CosignerSignature": {
        "type": "object",
        "description": "Cosigner signature entry for delta proposals",
        "required": [
          "signature",
          "timestamp",
          "signer_id"
        ],
        "properties": {
          "signature": {
            "$ref": "#/components/schemas/ProposalSignature"
          },
          "signer_id": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "CounterpartySummary": {
        "type": "object",
        "required": [
          "account_id",
          "direction"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/CounterpartyDirection"
          }
        }
      },
      "DashboardDeltaCategory": {
        "type": "string",
        "description": "Closed, stable enumeration of action categories. Adding a value is\na wire-contract change.",
        "enum": [
          "asset_transfer",
          "note_consumption",
          "note_creation",
          "account_storage_change",
          "guardian_switch",
          "custom"
        ]
      },
...
      "DeltaMetadata": {
        "type": "object",
        "description": "Persisted activity metadata for a canonical delta. Stored as JSONB\nin the `deltas.metadata` column. `None` for EVM deltas and any\nhistorical row never reprocessed by [`build_metadata`].",
        "required": [
          "category"
        ],
        "properties": {
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetSummary"
            },
            "description": "Assets surfaced in deterministic order. Empty when the\ntransaction does not move an asset or extraction failed.\nMulti-asset transactions populate every extractable entry so\nclients do not show a misleading single-asset summary."
          },
          "category": {
            "$ref": "#/components/schemas/DashboardDeltaCategory"
          },
          "counterparty": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CounterpartySummary",
                "description": "Counterparty of the transaction. `None` for transactions\nwithout a clear sender/recipient (admin ops, swaps, etc.)."
              }
            ]
          },
          "note_counts": {
            "$ref": "#/components/schemas/NoteCounts"
          },
          "proposal": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProposalMetadata",
                "description": "Multisig proposal intent lifted from the matching\n`delta_proposals` row at push time. Absent for single-key\n`push_delta`, EVM deltas, and pushes where no proposal matched."
              }
            ]
          }
        }
      },
      "DeltaObject": {
        "type": "object",
        "description": "Delta object",
        "required": [
          "account_id",
          "nonce",
          "prev_commitment",
          "delta_payload",
          "ack_sig",
          "ack_pubkey",
          "ack_scheme",
          "status"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "ack_pubkey": {
            "type": "string"
          },
          "ack_scheme": {
            "type": "string"
          },
          "ack_sig": {
            "type": "string"
          },
          "delta_payload": {
            "type": "object",
            "description": "Opaque, schema-free JSON payload describing the state delta."
          },
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeltaMetadata",
                "description": "Typed dashboard metadata derived at push time. Stored as JSONB\nin the `deltas.metadata` column. `None` for EVM deltas and any\nhistorical row never reprocessed by the push-time pipeline."
              }
            ]
          },
          "new_commitment": {
            "type": [
              "string",
              "null"
            ]
          },
          "nonce": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "prev_commitment": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DeltaStatus"
          }
        }
      },
...
      "DeltaStatus": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "timestamp",
              "proposer_id",
              "cosigner_sigs",
              "status"
            ],
            "properties": {
              "cosigner_sigs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CosignerSignature"
                }
              },
              "proposer_id": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "pending"
                ]
              },
              "timestamp": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "timestamp",
              "status"
            ],
            "properties": {
              "abandon_confirm_count": {
                "type": "integer",
                "format": "int32",
                "description": "Consecutive worker ticks that observed the on-chain commitment\nstill at this candidate's base after the abandon was requested.\nReset on a divergent observation, so only an unbroken streak\ncounts — the same stale-RPC shield as `divergence_count`.",
                "minimum": 0
              },
              "abandon_requested_at": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "RFC 3339 UTC timestamp of the client's abandon request\n(issue #319). While set, the status remains `candidate` — the\naccount stays locked — until the canonicalization worker\nresolves the intent after the abandon quarantine."
              },
              "divergence_count": {
                "type": "integer",
                "format": "int32",
                "description": "Consecutive canonicalization ticks that observed the on-chain\ncommitment at neither this delta's `prev_commitment` nor its\nexpected new commitment — i.e. the account advanced past the\nstate this candidate was built on. Reset to zero whenever a\nread shows the account still at the candidate's base, so only\nan unbroken streak counts. Requiring more than one observation\nbefore discarding shields against stale RPC reads.",
                "minimum": 0
              },
              "retry_count": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              "status": {
                "type": "string",
                "enum": [
                  "candidate"
                ]
              },
              "timestamp": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "timestamp",
              "status"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "canonical"
                ]
              },
              "timestamp": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "A candidate the worker gave up on, kept for background\nreconciliation (issue #345) instead of being deleted. A give-up\nverdict — retry exhaustion or confirmed divergence — is an\n*observation*, not proof the delta is wrong, so the worker keeps\nre-checking `stored base + delta` against the chain and promotes\nthe row if they ever match. Does NOT hold the pending-candidate\nlock: new submissions stay unblocked and supersede a retained\nrow at the same nonce. `timestamp` is when retention began and\nanchors the TTL after which the row is dropped for good.",
            "required": [
              "timestamp",
              "status"
            ],
            "properties": {
              "reason": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RetainReason"
                  }
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "retained"
                ]
              },
              "timestamp": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "timestamp",
              "status"
            ],
            "properties": {
              "reason": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/DiscardReason"
                  }
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "discarded"
                ]
              },
              "timestamp": {
                "type": "string"
              }
            }
          }
        ],
        "description": "Delta status state machine"
      },
...
      "MidenNetworkType": {
        "type": "string",
        "enum": [
          "local",
          "devnet",
          "testnet"
        ]
      },
      "NetworkConfig": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "network_type",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "miden"
                ]
              },
              "network_type": {
                "$ref": "#/components/schemas/MidenNetworkType"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "chain_id",
              "account_address",
              "multisig_validator_address",
              "kind"
            ],
            "properties": {
              "account_address": {
                "type": "string"
              },
              "chain_id": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "kind": {
                "type": "string",
                "enum": [
                  "evm"
                ]
              },
              "multisig_validator_address": {
                "type": "string"
              }
            }
          }
        ]
      },
      "NoteCounts": {
        "type": "object",
        "properties": {
          "input": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "output": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "x-pubkey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-pubkey"
      },
      "x-signature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-signature"
      },
      "x-timestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "x-timestamp"
      }
    }
  },
...
}